Reformat TelegramBot, make new BotService, deleted TelegramBotSetButtons.java, added method findUsersById to UserService
