+++ src/main/java/com/telegrambot/jd501/service/UserService.java
@@ -29,7 +29,7 @@ public class UserService {
      *
      * @return collection of User
      */
-    public Collection <User> getAllUser() {
+    public Collection<User> getAllUser() {
         return userRepository.findAll();
     }
 
@@ -116,8 +116,18 @@ public class UserService {
      *
      * @return List<User>
      */
-    public List<User> findUsersByAdoptedIsTrue () {
+    public List<User> findUsersByAdoptedIsTrue() {
         return userRepository.findUsersByIsAdoptedIsTrue();
+    }
+
+    /**
+     * find user if he exists by his ID
+     * Use method User repository {@link UserRepository#existsById(Object)}
+     *
+     * @return boolean
+     */
+    public boolean findUsersById(long userId) {
+        return userRepository.existsById(userId);
     }
